upd: rate status from int to float

+++ internal/models/rate.go
@@ -2,7 +2,7 @@ package models
 
 type RateCreate struct {
 	Title       string      `json:"title"`
-	Profit      int         `json:"profit"`
+	Profit      float64     `json:"profit"`
 	MinLockDays int         `json:"min_lock_days"`
 	Commission  int         `json:"commission"`
 	Properties  interface{} `json:"properties"`
